+++ backend/api/constant.go
@@ -1,4 +1,4 @@
-package magitrickleAPI
+package api
 
 const (
 	SocketPath = "/opt/var/run/magitrickle.sock"
