+++ backend/models/config/rule.go
@@ -1,7 +1,7 @@
 package config
 
 import (
-	"magitrickle/pkg/magitrickle-api/types"
+	"magitrickle/api/types"
 )
 
 type Rule struct {
