+++ backend/models/group.go
@@ -1,7 +1,7 @@
 package models
 
 import (
-	"magitrickle/pkg/magitrickle-api/types"
+	"magitrickle/api/types"
 )
 
 type Group struct {
