+++ backend/models/rule.go
@@ -4,7 +4,7 @@ import (
 	"regexp"
 	"strings"
 
-	"magitrickle/pkg/magitrickle-api/types"
+	"magitrickle/api/types"
 
 	"github.com/IGLOU-EU/go-wildcard/v2"
 )
